[add] insert location data list to ROOM database when stopRecording.

+++ app/src/main/java/ymse3p/app/audiorecorder/models/GpsData.kt
@@ -6,5 +6,5 @@ data class GpsData(
     val altitude: Double = 0.0,
     val bearing: Float = 0F,
     val speed: Float = 0F,
-    val time: Float = 0F,
+    val time: Long = 0,
 )
